Removed all dependencies to request and UI from PortletFinderUtil
PortletFinderUtil had all of its dependencies to the portlet request and
its UI removed. The search container is now entirely managed in the JSP
and portlet layers.

This will ease the usage of PortletFinderUtil in contexts not
necessarily related to rendering.

Some unused functions were also removed and most of the
PortletFinderUtil API has been made private to expose only the useful
part of the API.